Makes circle_id, username, picture_url optional
Makes the circle\_id, username, and picture\_url fields optional for both the User model and the Circle model.

This allows for greater flexibility in user and circle creation, as these fields are not always required upfront.

WIP on users POST request. Currently able to add test user to db.