Adds circle creation and retrieval
Implements the HTTP logic for creating and retrieving
circles, including input validation and sanitization.
Also updates the API documentation in README.

Still need to finish editing and removal for circles.

NEED TO VALIDATE AND SANITIZE REQUEST ID PARAMETER VALUES BEFORE THEY GET SENT TO THE DB IN USERS AND CIRCLES