Adds Google OAuth-based authentication
Implements user authentication using Google OAuth.

This introduces the `AuthController` to handle the OAuth flow,
registers the new controller to the application's routes,
and adds the `Imperial` package for OAuth integration.
Also removes some unneeded prints from the `CalendarEventsController`.

Able to access google oauth panel, but the callback is not working.

Need to attach protection to routes.